+++ dd-smoke-tests/iast-util/src/main/java/datadog/smoketest/springboot/controller/IastViewController.java
@@ -1,4 +1,4 @@
-package datadog.smoketest.springboot.controllers;
+package datadog.smoketest.springboot.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
