+++ dd-smoke-tests/iast-util/src/main/java/datadog/smoketest/springboot/controller/SimpleIastController.java
@@ -1,4 +1,4 @@
-package datadog.smoketest.springboot.controllers;
+package datadog.smoketest.springboot.controller;
 
 import java.io.PrintWriter;
 import java.util.Map;
