+++ dd-smoke-tests/spring-boot-2.6-webmvc/src/main/java/datadog/smoketest/springboot/SpringbootApplication.java
@@ -1,6 +1,6 @@
 package datadog.smoketest.springboot;
 
-import datadog.smoketest.springboot.controllers.SimpleIastController;
+import datadog.smoketest.springboot.controller.SimpleIastController;
 import java.util.Collections;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -19,7 +19,7 @@ import org.springframework.web.util.UrlPathHelper;
 public class SpringbootApplication {
 
   @Configuration
-  @ComponentScan(basePackages = {"datadog.smoketest.springboot.controllers"})
+  @ComponentScan(basePackages = {"datadog.smoketest.springboot.controller"})
   public static class WebConfig extends WebMvcConfigurerAdapter {
 
     @Override
